Final Version 1.0.0.1

+++ src/main/java/gs/sidartatech/safecheckout/Main.java
@@ -6,6 +6,7 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -25,6 +26,7 @@ import org.jutils.jprocesses.JProcesses;
 
 import gs.sidartatech.safecheckout.api.AppVersionAPI;
 import gs.sidartatech.safecheckout.api.AuthenticationAPI;
+import gs.sidartatech.safecheckout.domain.App;
 import gs.sidartatech.safecheckout.domain.AppVersion;
 import gs.sidartatech.safecheckout.domain.ControlExecution;
 import gs.sidartatech.safecheckout.domain.Routine;
@@ -38,8 +40,9 @@ import gs.sidartatech.safecheckout.service.ServiceFactory;
 
 public class Main {
 	private static Logger logger = Logger.getLogger(Main.class.getName());
-	public static final String API_URL = "http://127.0.0.1:9797/safe-checkout/api/";
+	public static final String API_URL = "http://191.252.222.32:9797/safe-checkout/api/";
 	private static final Integer ROUTINE_ID = 1;
+	private static final String VERSION = "1.0.0.1";
 	public static String token = "";
 	public static String versionPath;
 	public static String appPath;
@@ -63,6 +66,17 @@ public class Main {
 	        fh.setFormatter(sf);
 	        Logger.getLogger("").addHandler(fh);
 	        logger.log(Level.INFO, "Starting Safe-CheckOut");
+	        
+	        logger.log(Level.INFO, "Remove File Version...");
+			File versionFile = new File(versionPath + "/get-version.txt");
+			versionFile.delete();
+			logger.log(Level.INFO, "Remove File Version...(OK)");
+			logger.log(Level.INFO, "Create new File Version...");
+			versionFile.createNewFile();
+			FileWriter fw = new FileWriter(versionFile);
+			fw.write(VERSION);
+			fw.close();
+			logger.log(Level.INFO, "Create new File Version...(OK)");
 	        
 	        routineService = ServiceFactory.getInstance().getRoutineService();
 	        Routine routine = routineService.getByKey(ROUTINE_ID);
@@ -90,7 +104,7 @@ public class Main {
 	        File[] files = new File(versionPath).listFiles();
 	        if(files!=null) {
 	            for(File f: files) {
-	                if(!f.isDirectory() && !f.getName().equals("version.txt")) {
+	                if(!f.isDirectory() && !f.getName().equals("version.txt") && !f.getName().equals("get-version.txt") && !f.getName().equals("send-sale.txt")) {
 	                    f.delete();
 	                }
 	            }
@@ -99,7 +113,7 @@ public class Main {
 	        
 	        AppService appService = ServiceFactory.getInstance().getAppService();
 	        AppVersionService appVersionService = ServiceFactory.getInstance().getAppVersionService();
-	       
+	        
 	        for (AppVersion appVersion: appVersions) {
 	        	if (appService.getByKey(appVersion.getApp().getAppId()) != null) {
 					appService.update(appVersion.getApp());
@@ -112,6 +126,57 @@ public class Main {
 					appVersionService.insert(appVersion);
 				}
 	        }
+	        
+	        for (App app: appService.getAll()) {
+	        	if (app.getReplaceNow().equals("Y")) {
+	        		String appVersionFile = path + "/version/" + app.getFileName().split("\\.")[0] + ".txt";
+	        		if(Files.exists(Paths.get(appVersionFile))) {
+	        			try(BufferedReader br = new BufferedReader(new FileReader(new File(appVersionFile)))){
+		        			String currentVersion = "";
+		    				String line = "";
+		    				while((line = br.readLine()) != null) {
+		    					if (line != null) {
+		    						currentVersion = line.trim();
+		    					}
+		    				}
+		    				AppVersion appVersion = appVersionService.getLastVersion(app.getAppId());
+							if (!appVersion.getVersionNumber().equals(currentVersion)) {
+								if (app.getAppId().equals(3)) {
+									String changeVersionFile = path + "/version/change-version-file.txt";
+									Files.createFile(Paths.get(changeVersionFile));
+									FileWriter fileWriter = new FileWriter(changeVersionFile);
+									fileWriter.write(appVersion.getVersionPath());
+									fileWriter.close();
+								} else {
+									Files.delete(Paths.get(path + "/" + app.getFileName()));
+									File appFileVersion = new File(appVersion.getVersionPath());
+						    	    File newAppFileVersion = new File(path + "/" + app.getFileName());
+						    		
+						    	    InputStream fis = new FileInputStream(appFileVersion);
+						    	    OutputStream os = new FileOutputStream(newAppFileVersion);
+						        	
+						    	    byte[] buffer = new byte[1024];
+						    	    int length; 
+						    	    while ((length = fis.read(buffer)) > 0){
+						    	    	os.write(buffer, 0, length);
+						    	    }
+						    	    fis.close();
+						    	    os.close();
+								}
+							}
+		        		} catch (FileNotFoundException e) {
+		    				endOK = false;
+		    				messageError = e.getMessage();
+		    				logger.log(Level.SEVERE, e.getMessage(), e);
+		    			} catch (IOException e) {
+		    				endOK = false;
+		    				messageError = e.getMessage();
+		    				logger.log(Level.SEVERE, e.getMessage(), e);
+		    			}
+	        		}
+	        	}
+	        }
+	        
 	        logger.log(Level.INFO, "Update Versions...(OK)");
 	        
 	        logger.log(Level.INFO, "Checking Current Version...");
